:art: 0130
0130


Former-commit-id: 3f40e87342260463a433b06301f40cbedfcf5ef1

+++ BE/blur/blur-user-service/src/main/java/com/blur/userservice/global/security/LoginAuthenticationFilter.java
@@ -74,7 +74,7 @@ public class LoginAuthenticationFilter extends UsernamePasswordAuthenticationFil
         Date expiredTime = jwtTokenProvider.getExpiredTime(accessToken);
         String refreshToken = jwtTokenProvider.createJwtRefreshToken();
 
-        refreshTokenServiceImpl.updateRefreshToken(Long.valueOf(userId), jwtTokenProvider.getRefreshTokenId(refreshToken));
+        refreshTokenServiceImpl.updateRefreshToken(Integer.valueOf(userId), jwtTokenProvider.getRefreshTokenId(refreshToken));
 
         // 쿠키 설정
         ResponseCookie refreshTokenCookie = cookieProvider.createRefreshTokenCookie(refreshToken);
